requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
numpy==1.26.4
packaging==23.2
Pillow==10.1.0
pyTelegramBotAPI==4.15.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
telebot==0.0.5
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.9.0
urllib3==2.1.0
